nvidia/multi-agent-chatbot/assets/flux-service/requirements.txt
diffusers>=0.32.0
transformers>=4.45.0
accelerate
safetensors
fastapi
uvicorn[standard]
huggingface-hub>=0.27.0
pillow>=11.0.0
pydantic
numpy
onnxruntime-gpu>=1.19.2
sentencepiece
bitsandbytes
